Backend add Log

+++ Red_Social_Proyecto/Red_Social_Proyecto/Services/PublicationService.cs
@@ -7,6 +7,7 @@ using Red_Social_Proyecto.Dtos.Task;
 using Red_Social_Proyecto.Dtos.ValidationsDto;
 using Red_Social_Proyecto.Entities;
 using Red_Social_Proyecto.Services.Interfaces;
+using Red_Social_Proyecto.Services.LogsService.Interface;
 using Red_Social_Proyecto.SignalRConnect;
 
 namespace Red_Social_Proyecto.Services
@@ -16,15 +17,18 @@ namespace Red_Social_Proyecto.Services
         private readonly TodoListDBContext _context;
         private readonly IMapper _mapper;
         private readonly IHubContext<PublicationHub> _hubcontext;
+        private readonly ILogService _logService;
         private readonly HttpContext _httpContext;
         private readonly string _USER_ID;
 
         public PublicationService(TodoListDBContext context, IMapper mapper, 
-            IHttpContextAccessor httpContextAccessor, IHubContext<PublicationHub> hubcontext)
+            IHttpContextAccessor httpContextAccessor, IHubContext<PublicationHub> hubcontext,
+            ILogService logService)
         {
             _context = context;
             _mapper = mapper;
             _hubcontext = hubcontext;
+            _logService = logService;
             _httpContext = httpContextAccessor.HttpContext!; //para mandar el id del usuario que a creado la tarea
             var idClaim = _httpContext.User.Claims.
                 Where(x => x.Type == "UserId"). //Revertir el token y obtener el id del usuario
@@ -42,6 +46,8 @@ namespace Red_Social_Proyecto.Services
 
             _context.Publications.Add(publicationEntity);
             await _context.SaveChangesAsync();
+
+            await _logService.LogActionAsync("Creaste una publicación");
 
             var publicationDto = _mapper.Map<PublicationDto>(publicationEntity);
 
@@ -84,9 +90,7 @@ namespace Red_Social_Proyecto.Services
 
         public async Task<ResponseDto<bool>> DeletePublicationAsync(Guid publicationId)
         {
-
             var publication = await _context.Publications.FindAsync(publicationId);
-            publication.UserId = _USER_ID;
             if (publication == null)
             {
                 return new ResponseDto<bool>
@@ -98,8 +102,22 @@ namespace Red_Social_Proyecto.Services
                 };
             }
 
+            // Comprobar si el usuario actual es el dueño de la publicación
+            if (publication.UserId != _USER_ID)
+            {
+                return new ResponseDto<bool>
+                {
+                    Status = false,
+                    StatusCode = 403, // Código de estado para "Forbidden"
+                    Message = "No autorizado para eliminar esta publicación",
+                    Data = false
+                };
+            }
+
             _context.Publications.Remove(publication);
             await _context.SaveChangesAsync();
+
+            await _logService.LogActionAsync("Eliminaste una publicación");
 
             return new ResponseDto<bool>
             {
@@ -109,6 +127,35 @@ namespace Red_Social_Proyecto.Services
                 Data = true
             };
         }
+
+
+        public async Task<List<PublicationDto>> GetPublicationsForUserAndFollowersAsync(string userId)
+        {
+            var userPublications = await _context.Publications
+                .Where(p => p.UserId == userId)
+                .Include(p => p.User) // Incluye detalles del usuario
+                .ToListAsync();
+            
+
+            var followedUsersIds = await _context.Follows
+                .Where(f => f.FollowerId == userId)
+                .Select(f => f.FollowedId)
+                .ToListAsync();
+
+            var followedPublications = await _context.Publications
+                .Where(p => followedUsersIds.Contains(p.UserId))
+                .Include(p => p.User) // Incluye detalles del usuario
+                .ToListAsync();
+
+            // Combinar las publicaciones del usuario y las de los usuarios seguidos
+            var allPublications = userPublications.Concat(followedPublications).ToList();
+
+            
+
+            return _mapper.Map<List<PublicationDto>>(allPublications);
+        }
+
+
 
     }
 }
